fix: prevent home button overlap with sidebar
- Moved desktop sidebar from top-0 to top-16 to avoid overlapping home button
- Kept home button positioned at top-right (correct for RTL layout)
- Maintained proper z-index hierarchy (home button z-40, sidebar z-20)
- Fixed Layout.tsx file corruption from previous edit
- Home button now clearly visible and accessible on both desktop and mobile

Resolves home button visibility issue on desktop view.
